+++ lib/rubocop/formatter/formatter_set.rb
@@ -22,7 +22,8 @@ module RuboCop
         'worst'       => WorstOffendersFormatter,
         'tap'         => TapFormatter,
         'quiet'       => QuietFormatter,
-        'autogenconf' => AutoGenConfigFormatter
+        'autogenconf' => AutoGenConfigFormatter,
+        'pacman'      => PacmanFormatter
       }.freeze
 
       FORMATTER_APIS = %i[started finished].freeze